Aula41/requirements.txt
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.11.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
PyMySQL==0.9.3
PyMysqlDB==0.0.2
pytz==2019.3
six==1.14.0
Werkzeug==0.16.1
